+++ lib/cogment/client_actor.h
@@ -38,9 +38,9 @@ class Client_actor : public Actor {
 
   Future<void> init() override;
 
-  void dispatch_observation(const cogment::Observation& obs, bool end_of_trial) override;
-  void dispatch_reward(const ::cogment::Reward& reward) override;
-  void dispatch_message(int tick_id, const ::cogment::Message& message) override;
+  void dispatch_observation(cogment::Observation&& obs, bool end_of_trial) override;
+  void dispatch_reward(cogment::Reward&& reward) override;
+  void dispatch_message(cogment::Message&& message) override;
 
   bool is_active() const override;