+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Changed
+
+- Tick ID management centralized in orchestrator
+
 ## v1.0.0-alpha6 - 2021-03-10
 
 ### Added